requirements.txt
graphics.py==5.0.1.post1
numpy==2.1.1
